chore(backend): ensure that the validation process is handled in new created transaction
Before this commit I've noticed that the same transaction is used for every
request or at last for two concurrent requests.
When there are two request handling concurrently and one of them failed,
the second also failed too.
Because they are all handled in the same
transaction.

So, why it is the same transaction ?
Because Medusa all services are singleton, that means it is the same instance used
from the apps starts.
And the `atomicPhase_` only create new transaction manager if there no transaction manager attached
to the service.

https://github.com/medusajs/medusa/blob/3bf6e50744b56bb13903ff2e89ec83a798f439db/packages/medusa/src/loaders/plugins.ts#L226-L227

https://github.com/medusajs/medusa/issues/1834

In this commit we explicitly create a new transaction before
starting the validation