chore(test): create test for race condition testing
This test is about to ensure a good management of the race condition.
By race condition, we mean what will happen is two clients try to complete
their carts on the same time while the remaining quantities can not satisfy
the both orders.
The expected behavior suggest that the first client whose request is
handled first will be able to complete his cart by the second couldn't.
The first client is named Bob and the second Alice.